Move RelativeTo into but_rebase crate
This type really ought to be owned by the rebase engine. As such, I’ve moved the origional `RebaseTo` to be `RebaseToRef` and moved this owned version to be `RebaseTo`
