+++ crates/but-api/src/commit/types.rs
@@ -1,7 +1,6 @@
 use std::collections::BTreeMap;
 
 use but_core::{DiffSpec, tree::create_tree::RejectionReason};
-use but_rebase::graph_rebase::{Editor, ToSelector};
 
 /// Outcome after creating a commit.
 pub struct CommitCreateResult {
@@ -39,35 +38,4 @@ pub struct CommitInsertBlankResult {
     pub new_commit: gix::ObjectId,
     /// Commits that were replaced by this operation. Maps `old_id -> new_id`.
     pub replaced_commits: BTreeMap<gix::ObjectId, gix::ObjectId>,
-}
-
-/// Specifies a location relative to which a commit operation should occur.
-/// This is the fully-owned cousin of [but_rebase::graph_rebase::mutate::RelativeTo]
-/// as the [`but_api_macros::but_api`] macro doesn't support contained references or referenced parameters
-/// beyond a few well-known ones.
-#[derive(Debug, Clone)]
-pub enum RelativeTo {
-    /// Relative to a commit.
-    Commit(gix::ObjectId),
-    /// Relative to a reference.
-    Reference(gix::refs::FullName),
-}
-
-impl From<super::json::RelativeTo> for RelativeTo {
-    fn from(value: super::json::RelativeTo) -> Self {
-        match value {
-            super::json::RelativeTo::Commit(commit) => Self::Commit(commit),
-            super::json::RelativeTo::Reference(reference)
-            | super::json::RelativeTo::ReferenceBytes(reference) => Self::Reference(reference),
-        }
-    }
-}
-
-impl ToSelector for RelativeTo {
-    fn to_selector(&self, editor: &Editor) -> anyhow::Result<but_rebase::graph_rebase::Selector> {
-        match self {
-            Self::Commit(commit) => editor.select_commit(*commit),
-            Self::Reference(reference) => editor.select_reference(reference.as_ref()),
-        }
-    }
 }
