+++ crates/but-workspace/tests/workspace/commit/commit_create.rs
@@ -2,7 +2,7 @@ use anyhow::Result;
 use but_core::DiffSpec;
 use but_rebase::graph_rebase::{
     GraphExt as _, LookupStep as _,
-    mutate::{InsertSide, RelativeTo},
+    mutate::{InsertSide, RelativeToRef},
 };
 use but_workspace::commit::commit_create;
 
@@ -31,7 +31,7 @@ fn commit_above_commit() -> Result<()> {
     let outcome = commit_create(
         editor,
         worktree_changes_as_specs(&repo)?,
-        RelativeTo::Commit(two_id),
+        RelativeToRef::Commit(two_id),
         InsertSide::Above,
         "insert above commit",
         0,
@@ -77,7 +77,7 @@ fn commit_below_commit() -> Result<()> {
     let outcome = commit_create(
         editor,
         worktree_changes_as_specs(&repo)?,
-        RelativeTo::Commit(two_id),
+        RelativeToRef::Commit(two_id),
         InsertSide::Below,
         "insert below commit",
         0,
@@ -117,7 +117,7 @@ fn commit_above_reference() -> Result<()> {
     let outcome = commit_create(
         editor,
         worktree_changes_as_specs(&repo)?,
-        RelativeTo::Reference(reference.name()),
+        RelativeToRef::Reference(reference.name()),
         InsertSide::Above,
         "insert above reference",
         0,
@@ -168,7 +168,7 @@ fn commit_below_merge_commit_uses_first_parent() -> Result<()> {
     let outcome = commit_create(
         editor,
         worktree_changes_as_specs(&repo)?,
-        RelativeTo::Commit(merge_id),
+        RelativeToRef::Commit(merge_id),
         InsertSide::Below,
         "insert below merge",
         0,
@@ -210,7 +210,7 @@ fn commit_all_rejected_is_noop() -> Result<()> {
             path: "does-not-exist".into(),
             hunk_headers: vec![],
         }],
-        RelativeTo::Commit(two_id),
+        RelativeToRef::Commit(two_id),
         InsertSide::Above,
         "no-op commit",
         0,
